% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider-package.R
\docType{data}
\name{anoteropsis}
\alias{anoteropsis}
\title{Cytochrome oxidase I (COI) sequences of New Zealand _Anoteropsis_ species}
\format{A DNAbin object containing 33 sequences with a length of 409 base
pairs stored as a matrix.}
\source{
Vink, C. J., and Paterson, A. M. (2003). Combined molecular and
morphological phylogenetic analyses of the New Zealand wolf spider genus
_Anoteropsis_ (Araneae: Lycosidae). _Molecular Phylogenetics and Evolution_
*28* 576-587.
}
\description{
A set of 33 sequences of the mitochondrial protein-coding gene cytochrome
oxidase I from 20 species of the New Zealand wolf spider genus
\emph{Anoteropsis} (Lycosidae) and two species of \emph{Artoria} as
outgroups. The sequences are available on GenBank as accession numbers
AY059961 through AY059993.
}
\keyword{Datasets}
