\name{dolomedes}
\docType{data}
\alias{dolomedes}
\title{Cytochrome oxidase I (COI) sequences of New Zealand \emph{Dolomedes} species}

\description{This is a set of 33 sequences of the mitochondrial protein-coding gene cytochrome oxidase I from 20 species of the New Zealand spider genus \emph{Dolomedes} and two species of \emph{Artoria} as outgroups.}

\usage{dolomedes}

\format{A DNAbin object containing 33 sequences with a length of 409 base pairs}