% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider-package.R
\docType{data}
\name{sarkar}
\alias{sarkar}
\title{Dummy sequences illustrating the categories of diagnostic nucleotides}
\format{
A DNAbin object containing 8 sequences with a length of 18 base
pairs stored as a matrix.
}
\source{
Sarkar, I., Planet, P., & DeSalle, R. (2008). CAOS software for use
in character- based DNA barcoding. _Molecular Ecology Resources_ *8*
1256-1259
}
\description{
A set of 8 dummy sequences published in Sarkar et al 2008 to illustrate the
different categories of diagnostic nucleotides.
}
\keyword{Datasets}
