% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeBarcode.R
\name{seeBarcode}
\alias{seeBarcode}
\title{Create illustrative barcodes}
\usage{
seeBarcode(seq, col = c("green", "blue", "black", "red"))
}
\arguments{
\item{seq}{A single sequence of class `DNAbin'.}

\item{col}{A character vector of length 4 giving colours to represent A, G,
C and T respectively.}
}
\value{
Plots an illustrative barcode.
}
\description{
This function plots an illustrative barcode consisting of vertical bands in
four colours corresponding to the DNA bases adenine (A), cytosine (C),
guanine (G) and thiamine (T).
}
\details{
Green, blue, black and red are the standard colours representing A, G, C and
T respectively.
}
\examples{

layout(matrix(1:6, ncol=1))
par(mar=c(0.5, 0, 0.5, 0))
data(woodmouse)
seeBarcode(woodmouse[1,])
seeBarcode(woodmouse[1,], col=c("pink", "orange", "steelblue", "yellow"))
seeBarcode(woodmouse[1,], col=c("black", "white", "white", "black"))
apply(woodmouse[1:3,], MARGIN=1, FUN=seeBarcode)

}
\author{
Samuel Brown <s_d_j_brown@hotmail.com>
}
\keyword{Barcoding}
\keyword{Utilities}
