% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIDER.R
\docType{package}
\name{SPIDER}
\alias{SPIDER}
\title{SPIDER (surface protein prediction using deep ensembles from single-cell RNA-seq)}
\description{
A context-agnostic zero-shot deep ensemble model which enables the large-scale prediction of cell surface protein abundance.
}
\references{
Ruoqiao Chen, Jiayu Zhou, Bin Chen
[https://github.com/Bin-Chen-Lab/spider](https://github.com/Bin-Chen-Lab/spider)
}
