% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIDER_predict.R
\name{SPIDER_predict}
\alias{SPIDER_predict}
\title{Surface protein abundance prediction}
\usage{
SPIDER_predict(
  seurat_data,
  tissue = "pancreas",
  disease = "healthy",
  SPIDER_model_file_path,
  use_cell_type = "SingleR",
  query_cell_type = NULL,
  protein = "All",
  use_pretrain = "T",
  save_path,
  use_python_path,
  scarches_path,
  all_trainable_proteins_gene_names = NULL,
  file_A_B_C_matching_table = NULL
)
}
\arguments{
\item{seurat_data}{Query transcriptomes after prepocessing. Including Seurat log normalization, clustering and umap reductions. You must also use seurat_data[["study"]] = ... to specify the batch IDs for all cells.}

\item{tissue}{The tissue source of the query transcriptomes. Default options include: 'bone marrow', 'brain', 'blood', 'pleura', 'peritoneum'.}

\item{disease}{The disease state of the query transcriptomes. Default options include: 'healthy', 'mesothelioma', 'glioblastoma', 'leukemia'.}

\item{SPIDER_model_file_path}{The path to the saved pretrained SPIDER model weights.}

\item{use_cell_type}{The method name of cell type annotations. Default option is 'SingleR'.}

\item{query_cell_type}{A data frame specifying cell types for the query dataset. The row names should be cell IDs and there should be one column named 'final_celltype' specifying your defined cell type for every cell. This parameter is only needed if users do not set use_cell_type = 'SingleR'. Otherwise if users set use_cell_type = 'SingleR', they do not need to provide any data frame for the query_cell_type parameter. Default is NULL.}

\item{protein}{A vector specifying corresponding gene names of proteins to be predicted. Default option is 'All', where all surface proteins listed in UniProt will be predicted.}

\item{use_pretrain}{Whether to use pretrained SPIDER model weights or not. If yes, set use_pretrain = 'T', otherwise set use_pretrain = 'F'.}

\item{save_path}{The path to the directory where you want to save your prediction results.}

\item{use_python_path}{The path to the specific version of python for R reticulate. This parameter is only needed if you use a separate version of python for R reticulate from your default python configuration for reticulate. It will automatically pass this parameter to reticulate's "use_python" function. Otherwise just set this parameter to NULL.}

\item{scarches_path}{The path to the directory where the scArches package is downloaded.}

\item{all_trainable_proteins_gene_names}{A data frame matching all trainable proteins' (i.e., seen proteins in your reference set) names to their corresponding gene names. Columns must contain column names: 'consistent_protein_name' specifying the protein names, and 'gene_name' specifying the corresponding gene names. This parameter is only needed if users set use_pretrain = 'F', otherwise just set this parameter to NULL.}

\item{file_A_B_C_matching_table}{A data frame specifying the path to the reference data split by their context factors. Must containing column names: 'file_B' specifying all file paths of embedded reference transcriptomes, 'file_C' specifying all corresponding file paths of reference CLR normalized ADT counts, 'tissue_class' specifying corresponding reference tissue classes using numeric numbers (0, 1, 2, ...), 'disease_class' specifying corresponding reference disease classes using numeric numbers (0, 1, 2, ...), 'cell_type_class' specifying corresponding reference cell type classes using numeric numbers (0, 1, 2, ...). Make sure that each row only contain one cell type, one disease and one tissue correctly describing the states of cells in the corresponding transcriptome and ADT files. This parameter is only needed if users set use_pretrain = 'F', otherwise just set this parameter to NULL.}
}
\value{
Predicted surface protein abundance and confidence scores will be saved to your indicated directory save_path.
}
\description{
Predict surface protein abundance on provided query transcriptomes.
}
