\name{anoteropsis}
\docType{data}
\alias{anoteropsis}
\title{Cytochrome oxidase I (COI) sequences of New Zealand \emph{Anoteropsis} species}

\description{This is a set of 33 sequences of the mitochondrial protein-coding gene cytochrome oxidase I from 20 species of the New Zealand spider genus \emph{Anoteropsis} and two species of \emph{Artoria} as outgroups.}

\usage{anoteropsis}

\format{A DNAbin object containing 33 sequences with a length of 409 base pairs}