\name{dolomedes}
\docType{data}
\alias{dolomedes}
\title{Cytochrome oxidase I (COI) sequences of New Zealand _Dolomedes_ species}

\description{This is a set of 37 sequences of the mitochondrial protein-coding gene cytochrome oxidase I from the 4 New Zealand species of the spider genus \emph{Dolomedes}. These sequences are available on GenBank as accession numbers GQ337328 through GQ337385}

\usage{dolomedes}

\format{A DNAbin object containing 37 sequences with a length of 850 base pairs}

\source{
Vink, C. J., and Duperre, N. (2010). Pisauridae (Arachnida: Araneae). _Fauna of New Zealand_, *64* , 1-54.}

\keyword{Datasets}