\name{salticidae}
\docType{data}
\alias{salticidae}
\title{Cytochrome oxidase I (COI) sequences of world-wide species of Salticidae}

\description{A set of 41 sequences of the mitochondrial protein-coding gene cytochrome oxidase I from 41 species of the jumping spider family Salticidae.The sequences are available on GenBank as accession numbers AY297360 through AY297400.}

\usage{anoteropsis}

\format{A DNAbin object containing 41 sequences with a length of 409 base pairs stored as a list.}

\source{
Maddison, W. P., and Hedin, M. C. (2003). Jumping spider phylogeny (Araneae: Salticidae). _Invertebrate Systematics_ *17* 529-549.
}

\keyword{Datasets}