\name{spider-package}
\alias{spider-package}
\alias{spider}
\docType{package}
\title{
Species Identity and Evolution in R
}
\description{
A package for the analysis of species limits and DNA barcoding data
}
\details{
\tabular{ll}{
Package: \tab spider\cr
Type: \tab Package\cr
Version: \tab 1.0-5\cr
Date: \tab 2011-09-06\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Samuel Brown, Rupert Collins, Stephane Boyer, Marie-Caroline Lefort, Jagoba Malumbres-Olarte, Cor Vink, Rob Cruickshank

Maintainer:
Samuel Brown <s_d_j_brown@hotmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
\code{\link[ape:ape-package]{ape}} ~~
\code{\link[pegas:pegas-package]{pegas}}
}
\examples{
}
